fix: fp16_to_f32 type bug + add integration tests and driver benchmark
fp16_to_f32: result type was f32 instead of u32, causing
`result as u32` to do float-to-int cast instead of bitcast.
All fp16 conversions returned wrong values.

Added 13 integration tests: surface lifecycle, fp16 roundtrip,
MIL compilation, model lifecycle (compile/load/run/unload),
identity matmul correctness, error handling.

Added driver benchmark: surface create, MIL compile, load/unload,
dispatch overhead, fp16 throughput.

M1 Pro results:
  surface create:      0.03-0.07 ms
  MIL compile:         18-20 ms
  compile+load+unload: 23 ms
  dispatch overhead:   0.24 ms
  f32→fp16 16M:        39 GB/s
  fp16→f32 16M:        15 GB/s

+++ src/surface.rs
@@ -113,7 +113,7 @@ impl Drop for AneSurface {
 pub fn fp16_to_f32(v: u16) -> f32 {
     #[cfg(target_arch = "aarch64")]
     {
-        let result: f32;
+        let result: u32;
         unsafe {
             std::arch::asm!(
                 "fmov h0, {src:w}",
@@ -124,7 +124,7 @@ pub fn fp16_to_f32(v: u16) -> f32 {
                 out("v0") _,
             );
         }
-        f32::from_bits(result as u32)
+        f32::from_bits(result)
     }
     #[cfg(not(target_arch = "aarch64"))]
     {
